docs: Fix NPM example to use JS instead of TS syntax
The code example specifies that this is JavaScript code in the comment (index.js), but the code is actually only compatible with TypeScript due to the colon instead of equal sign. The first example above this one using the code snippit correctly uses the equal sign, only this NPM example uses the colon. I’ve simply updated the colon to be correct JS code and match the previous example. Alternatively we could update the comment to say `// index.ts`, but I think this is better for consistency sake as the previous example is in JS not TS.